test(product): split support sanitizer coverage
